rework digit int extensions into ulong extensions

+++ punku/Math/Math.cs
@@ -4,15 +4,15 @@ namespace Punku
 {
 	public static class Math
 	{
-		public static bool IsNarcissisticNumber (int n)
+		public static bool IsNarcissisticNumber (ulong n)
 		{
 			// NOTE: there can be a rounding error due to Math.Pow using doubles
 
-			int res = 0;
+			ulong res = 0;
 			int count = n.CountDigits ();
 
 			foreach (var x in n.Digits ())
-				res += (int)System.Math.Pow (x, count);
+				res += (ulong)System.Math.Pow (x, count);
 
 			if (n == res)
 				return true;
